Change Drawer interface methods to use float64.

+++ drawer.go
@@ -14,7 +14,7 @@ const (
 
 type Drawer interface {
 	// DrawCell draws a rune in a cell.
-	DrawCell(cell *Cell, column, row float32)
+	DrawCell(cell *Cell, column, row float64)
 
 	// DrawCellRect draws a rectangle of the desired color within
 	// a cell.
@@ -22,7 +22,7 @@ type Drawer interface {
 	// The rectangle described by rect dictates how much of the cell the solid
 	// block of color fills. 0,0 represents the top left of the cell, and 1,1
 	// the bottom right of the cell.
-	DrawCellRect(column, row float32, rect geometry.RectFloat, colour color.Color)
+	DrawCellRect(column, row float64, rect geometry.RectFloat, colour color.Color)
 
 	// DrawCellsRect draws a rectangle of solid colour spanning some
 	// or all of the cells.
